Prime, Prelevement et Credit __ Version Final
Prime, Prelevement et Credit __ Version Final

+++ employee_prelevement_model/prelevement.py
@@ -4,6 +4,7 @@ from odoo import fields, models, api
 from odoo.exceptions import ValidationError
 from math import *
 from calendar import monthrange
+from datetime import datetime
 
 class prelevement(models.Model):
     _name = "hr.prelevement"
@@ -22,13 +23,19 @@ class prelevement(models.Model):
         ("en_jour","Prélèvement En Jour"),
         ],"Type de prélèvement",
     )
-    type_prelevemenet_en_jour_nbrjour = fields.Float("Jours travaillés par mois")
+    type_prelevemenet_en_jour_nbrjour = fields.Float("Nombre de jours", default="1")
     type_prelevemenet_en_jour_salairejour = fields.Float("Salaire du jour")
     
     is_credit = fields.Boolean("Crédit?", default = False)
     
+    first_period_id = fields.Many2one("account.month.period", string = "Première Période", required=False)
+    echeance = fields.Monetary("Échéance", currency_field = "currency_id", required=False)
+
     @api.model
     def create(self, vals):
+        if vals.get("type_prelevement") and vals["type_prelevement"] == "en_jour":
+            periode_retournee = self.recuperer_id_periode(vals["date_fait"])
+            vals["first_period_id"] = periode_retournee
         return super(prelevement, self).create(vals)
          
     def write(self, vals):
@@ -40,12 +47,19 @@ class prelevement(models.Model):
             raise ValidationError("Le montant payé doit être inférieur du montant et strictement supérieur à 0.")
         if vals.get("state") and vals["state"] == "draft" and self.state == "annulee":
             for ligne in self.paiement_prelevement_ids:
+                if ligne.state == "paye":
+                    raise ValidationError("Erreur, Vous avez au moins une période payée, vous devez régler la situation.")
                 ligne.unlink()
         if vals.get("state") and vals["state"] == "validee":
-            if vals.get("echeance") or vals.get("montant_total_prime"):
-                for ligne in self.paiement_prelevement_ids:
-                    ligne.unlink()
+            for ligne in self.paiement_prelevement_ids:
+                if ligne.state == "paye":
+                    raise ValidationError("Erreur, Vous avez au moins une période payée, vous devez régler la situation.")
+                ligne.unlink()
             self.compute_prelevement()
+        if vals.get("state") and vals["state"] == "annulee":
+            for ligne in self.paiement_prelevement_ids:
+                if ligne.state == "paye":
+                    raise ValidationError("Erreur, Vous avez au moins une période payée, vous devez régler la situation.")
         return super(prelevement, self).write(vals)
     
 
@@ -54,8 +68,7 @@ class prelevement(models.Model):
             res = self.montant_total_prime / self.echeance
             nbr_periodes = ceil(res)            
             self.compute_alimenter_paiement_prelevement(nbr_periodes)
-
-
+            
 
     def compute_alimenter_paiement_prelevement(self,nbr_periodes):
         for rec in self:
@@ -153,20 +166,39 @@ class prelevement(models.Model):
             raise ValidationError("Erreur, Seulement les administrateurs et les agents de paie qui peuvent changer le status.")
         
 
-    @api.onchange("date_fait")
-    def onchange_date_fait_calcule(self):
+    def nbr_days_of_current_month(self):
         year_val = self.date_fait.year
         month_val = self.date_fait.month
         nbr_days = monthrange(year_val, month_val)[1]
-        self.type_prelevemenet_en_jour_nbrjour = nbr_days
+        return nbr_days
 
+    def recuperer_salaire(self):
+        profile_paie_employee = self.env['hr.employee'].browse(self.employee_id.id)
+        definition_nbr_jour = profile_paie_employee.pp_personnel_id_many2one.definition_nbre_jour_worked_par_mois
+        if definition_nbr_jour == "nbr_saisie":
+            self.type_prelevemenet_en_jour_salairejour = profile_paie_employee.pp_personnel_id_many2one.salaire_jour
+        elif definition_nbr_jour == "jr_mois":
+            nbr_days_of_current_month = self.nbr_days_of_current_month()
+            self.type_prelevemenet_en_jour_salairejour = profile_paie_employee.salaire_actuel / nbr_days_of_current_month
 
     @api.onchange("employee_id")
     def onchange_employee_nbr_jour(self):
-        profile_paie_employee = self.env['hr.employee'].browse(self.employee_id.id).pp_personnel_id_many2one
-        definition_nbr_jour = profile_paie_employee.definition_nbre_jour_worked_par_mois
-        print(definition_nbr_jour)
-        if definition_nbr_jour == "nbr_saisie":
-            self.type_prelevemenet_en_jour_salairejour = profile_paie_employee.salaire_jour
-        elif definition_nbr_jour == "jr_mois":
-            self.type_prelevemenet_en_jour_salairejour = self.employee_id.id.salaire_actuel 
+        self.recuperer_salaire()
+
+    @api.onchange("type_prelevemenet_en_jour_nbrjour","type_prelevemenet_en_jour_salairejour","type_prelevement")
+    def calculer_montant_onchange(self):
+        if self.type_prelevement == "en_jour" and self.is_credit != True:
+            nbr_jour = self.type_prelevemenet_en_jour_nbrjour
+            salaire_jr = self.type_prelevemenet_en_jour_salairejour
+            self.montant_total_prime = nbr_jour * salaire_jr
+            self.echeance = self.montant_total_prime
+
+    def recuperer_id_periode(self,date_fait):
+        query = """
+                SELECT id from account_month_period
+                WHERE '%s' between date_start and date_stop;
+            """  % (date_fait)
+        self.env.cr.execute(query)
+        res = self.env.cr.fetchall()
+        print(res[0][0])
+        return res[0][0]
